requirements.txt
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.7
django-filter==2.4.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
Faker==8.11.0
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.1
pandas==1.3.1
Pillow==8.2.0
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
